Moved property definition to its own file.
